Fix timetable reminders cancelled on navigation
