+++ components/Timetable.js
@@ -192,86 +192,10 @@ export default function Timetable({ role = "student" }) {
         if (isMounted()) setIsSyncing(false);
       }
     }
+    window.dispatchEvent(new Event("timetable-updated"));
   };
 
-  // Timetable push reminder scheduler
-  useEffect(() => {
-    if (pushStatus !== "granted" || !mounted) return;
-
-    const timerIds = [];
-    const todayName = new Date().toLocaleDateString("en-US", { weekday: "long" });
-    const todayClasses = timetableData[todayName] || [];
-
-    todayClasses.forEach((cls) => {
-      const [startStr] = cls.time.split("-");
-      if (!startStr) return;
-      const [hours, minutes] = startStr.split(":").map(Number);
-
-      const now = new Date();
-      const classTime = new Date();
-      classTime.setHours(hours, minutes, 0, 0);
-
-      // Reminder is 10 minutes before class
-      const reminderTime = new Date(classTime.getTime() - 10 * 60 * 1000);
-
-      if (classTime > now) {
-        if (reminderTime > now) {
-          const delay = reminderTime.getTime() - now.getTime();
-          const timerId = setTimeout(() => {
-            triggerNotification(cls);
-          }, delay);
-          timerIds.push(timerId);
-        } else {
-          // Class starts in less than 10m but hasn't started yet - trigger alert immediately
-          triggerNotification(cls, true);
-        }
-      }
-    });
-
-    return () => {
-      timerIds.forEach((id) => clearTimeout(id));
-    };
-  }, [pushStatus, timetableData, mounted]);
-
-  const triggerNotification = (cls, immediate = false) => {
-    if (typeof window === "undefined" || !("Notification" in window) || Notification.permission !== "granted") return;
-
-    const [startStr] = cls.time.split("-");
-    if (!startStr) return;
-    const [hours, minutes] = startStr.split(":").map(Number);
-    const classTime = new Date();
-    classTime.setHours(hours, minutes, 0, 0);
-
-    const minsLeft = immediate
-      ? Math.max(1, Math.round((classTime.getTime() - Date.now()) / 60000))
-      : 10;
-
-    const title = `Class starting in ${minsLeft}m: ${cls.subject}`;
-    const options = {
-      body: `📍 Location: ${cls.room}\n👨‍🏫 Instructor: ${cls.teacher}\n⏰ Schedule: ${cls.time}`,
-      icon: "/logo-icon.png",
-      badge: "/logo-icon.png",
-      vibrate: [100, 50, 100],
-      tag: `class-reminder-${cls.subject}-${cls.time}`,
-      data: {
-        url: "/timetable"
-      },
-      actions: [
-        { action: "open", title: "View Timetable" },
-        { action: "close", title: "Dismiss" }
-      ]
-    };
-
-    if ("serviceWorker" in navigator) {
-      navigator.serviceWorker.ready.then((registration) => {
-        registration.showNotification(title, options);
-      }).catch(() => {
-        new Notification(title, options);
-      });
-    } else {
-      new Notification(title, options);
-    }
-  };
+  // Timetable push reminder scheduler is now handled globally by useTimetableReminders hook
 
   const handleTogglePush = async () => {
     if (typeof window === "undefined" || !("Notification" in window)) return;
